+++ packages/openapi3-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # API Elements: OpenAPI 3 Parser Changelog
 
+## Unreleased
+
+### Enhancements
+
+- added a Link element to the specific format/version in the parse result.
+
 ## 0.15.0 (2020-08-06)
 
 ### Enhancements